refactor(config): leverage tagged enum types to allow selecting storage type without manually specifying config values
