+++ htsget-config/config.toml
@@ -1,44 +1,37 @@
-ticket_server_addr = '127.0.0.1:8080'
+ticket_server_addr = "127.0.0.1:8080"
 ticket_server_cors_allow_credentials = false
-ticket_server_cors_allow_origins = ['http://localhost:8080']
-ticket_server_cors_allow_headers = 'All'
-ticket_server_cors_allow_methods = 'All'
+ticket_server_cors_allow_origins = ["http://localhost:8080"]
+ticket_server_cors_allow_headers = "All"
+ticket_server_cors_allow_methods = "All"
 ticket_server_cors_max_age = 86400
 ticket_server_cors_expose_headers = []
-
-data_server = "None"
-#data_server_disabled = true
-data_server_addr = '127.0.0.1:8082'
-data_server_local_path = 'data'
-data_server_serve_at = '/data'
+data_server_enabled = true
+data_server_addr = "127.0.0.1:8081"
+data_server_local_path = "data"
+data_server_serve_at = "/data"
 data_server_cors_allow_credentials = false
-data_server_cors_allow_origins = ['http://localhost:8080']
-data_server_cors_allow_headers = 'All'
-data_server_cors_allow_methods = 'All'
+data_server_cors_allow_origins = ["http://localhost:8080"]
+data_server_cors_allow_headers = "All"
+data_server_cors_allow_methods = "All"
 data_server_cors_max_age = 86400
 data_server_cors_expose_headers = []
 
 [[resolvers]]
-regex = '.*'
-substitution_string = '$0'
+regex = ".*"
+substitution_string = "$0"
+storage = "S3"
 
-[resolvers.storage]
-scheme = 'Http'
-authority = '127.0.0.1:8081'
-local_path = 'data'
-path_prefix = '/data'
-
-[resolvers.guard]
-allow_reference_names = 'All'
-allow_fields = 'All'
-allow_tags = 'All'
+[resolvers.allow_guard]
+allow_reference_names = "All"
+allow_fields = "All"
+allow_tags = "All"
 allow_formats = [
-    'BAM',
-    'CRAM',
-    'VCF',
-    'BCF',
+    "BAM",
+    "CRAM",
+    "VCF",
+    "BCF",
 ]
 allow_classes = [
-    'body',
-    'header',
+    "body",
+    "header",
 ]
